// apps/desktop/src/ipc/generated/ProductionLocationCreateArgs.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ProductionLocationStatus } from "./ProductionLocationStatus";

export type ProductionLocationCreateArgs = { name: string, address?: string | null, contact?: string | null, status?: ProductionLocationStatus | null, };